Moved scheduler selection to before marking page as downloaded
Moved download scheduler to LinkStorage eventualy could cause lots
of links being maked as downloaded without actually downloading
when the crawler is stopped (links in the scheduler are already marked
as downloaded, but may not be downloaded if the crawler is stopped).
This commit alleviates that problem, makes the code a little bit
simpler and also removes synchronizations on the scheduler.